one more unneeded int

+++ src/coupling_scheme.jl
@@ -52,7 +52,7 @@ function possible_lsLS(k::Int, jpR::jp, jps::T where {T<:Vector{jp}})
     LSv = possible_ls(jpR, jps[k]; jp=jps[4])
     return [(ls=ls, LS=LS) for (ls, LS) in Iterators.product(lsv, LSv)]
 end
-function possible_lsLS(k::Int, two_s::Int, parity::Char, two_js, Ps)
+function possible_lsLS(k::Int, two_s, parity::Char, two_js, Ps)
     jpR = jp(two_s // 2, parity)
     jps = jp.(zip(Tuple(two_js) .// 2, Ps))
     return possible_lsLS(k, jpR, jps)
